test : ChessBoard 클래스 생성에 따른 테스트 이동
